+++ src/net/bitpot/railways/actions/CopyRoutePathAction.java
@@ -1,0 +1,12 @@
+package net.bitpot.railways.actions;
+
+import net.bitpot.railways.models.Route;
+import net.bitpot.railways.utils.RailwaysUtils;
+
+public class CopyRoutePathAction extends CopyRouteActionBase {
+
+    @Override
+    public String getRouteValue(Route route) {
+        return RailwaysUtils.stripRequestFormat(route.getPath());
+    }
+}