+++ CHANGES.md
@@ -4,6 +4,7 @@ Railways - Change Notes
 ### 0.8.4 (unreleased)
 
 * Add ability to copy route path
+* Add ability to copy data from multiple selected rows
 
 
 ### 0.8.3 (2015-05-16)